+++ frontend/src/pages/Courses/Courses.module.css
@@ -1,19 +1,10 @@
 
 @import url('https://fonts.googleapis.com/css2?family=Open+Sans:ital,wght@0,300..800;1,300..800&display=swap');
 
-/* UTIL */
-
-.row {
-    display: flex;
-    flex-direction: row;
-}
-
-.column {
+.Column {
     display: flex;
     flex-direction: column;
 }
-
-/* CONTAIN */
 
 .CoursesPage {
     position: absolute;
@@ -29,7 +20,7 @@
     padding-bottom: 200px;
 }
 
-.AddCourseButton {
+.EditButton {
     position: fixed;
 
     top: 95px;
@@ -50,356 +41,3 @@
 
     z-index: 2;
 }
-
-.AddCourseButton:active {
-    background-color: #9cccff;
-}
-
-.AddCourseMenuDormant {
-    position: fixed;
-
-    top: 95px; 
-    left: 20px;
-
-    width: 30px; 
-    height: 30px;
-
-    border-radius: 50%; 
-    transform: scale(0); 
-    transition: transform 0.5s, width 0.5s, height 0.5s, border-radius 0.5s;
-    background-color: white;
-    /* border: 1px solid #61ADFE; */
-    box-shadow: 0px 0px 10px rgba(0, 0, 0, 0.125);
-    z-index: 1; 
-}
-
-.AddCourseMenuActive {
-    top: 135px; 
-
-    width: 315px; 
-    height: 700px; 
-
-    border-radius: 15px; 
-    transform: scale(1); 
-
-    /* border: 1px solid blue; */
-}
-
-/* COURSES */
-
-.addCourseButton {
-    display: flex;
-    justify-content: center; /* Center content horizontally */
-    align-items: center; /* Center content vertically */
-
-    width: 36px;
-    height: 36px;
-
-    border-radius: 50%; /* Use 50% for a perfect circle */
-    margin-bottom: 5px;
-
-    background-color: #F5F5F5;
-
-    cursor: pointer; /* Change cursor on hover */
-}
-
-.addCourseButton:hover,
-.addCourseButton:active {
-    background-color: #E0E0E0; /* Lighter color when hovered or clicked */
-}
-
-
-.courseBox {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-
-    width: 425px;
-    height: 36px;
-
-    border-radius: 16px;
-    margin-bottom: 5px;
-
-    padding-left: 10px;
-    padding-right: 10px;
-
-    background-color: #F5F5F5;
-    transition: filter 0.4s ease;
-}
-
-.courseBox:hover {
-    cursor: pointer;
-    filter: brightness(95%);
-}
-
-.checkmark {
-    justify-content: center;
-    text-align: center;
-
-    font-weight: 550;
-    margin-right: 2px;
-}
-
-/* top top */
-
-.Grade {
-    font-weight: 600;
-    font-size: 25px;
-    margin-right: 10px;
-}
-
-
-
-
-.yearComponent {
-    display: flex;
-    flex-direction: column;
-    margin-bottom: 20px;
-    transition: transform 0.5s ease, height 0.5s ease, width 0.5s ease;
-  }
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-/* sem meta */
-
-.MetadataColumn {
-    font-size:small;
-    display: flex;
-    flex-direction: column;
-    margin-right: 4px;
-}
-
-.MetadataHeading {
-    color: #727272;
-    margin-bottom: 2px;
-}
-
-.evaluateBox {
-    display: inline-flex;
-    align-items: center;
-    justify-content: center;
-
-    width: max-content;
-    height: 20px;
-    padding: 0 6px; 
-
-    font-size: 11px;
-    font-weight: 510;
-
-    border-radius: 4px; 
-    color: #5aac08; 
-    background-color: #D9F4CF; 
-}
-
-.distBox {
-    display: inline-flex;
-    align-items: center;
-    justify-content: center;
-
-    width: max-content;
-    height: 20px;
-    padding: 0 6px; 
-
-    font-size: 12px;
-    font-weight: 510;
-
-    border-radius: 4px; 
-}
-
-.countBox {
-    display: inline-flex;
-    align-items: center;
-    justify-content: center;
-    margin-right: 10px;
-
-    width: max-content;
-    height: 18px;
-    padding: 0 6px; 
-
-    font-size: 12px;
-    font-weight: 500;
-
-    border: 1px solid grey; 
-    border-radius: 4px;
-    color: black;
-}
-
-
-
-.optionsButton {
-    background-color: white;
-    border: none;
-    border-radius: 8px;
-    color: #598FF4;
-    padding-right: 10px;
-    padding-left: 10px;
-    height: 30px;
-    font-size: 18px;
-    cursor: pointer;
-}
-  
-.optionsButton:hover {
-    background-color: #fafafa; /* Grey color */
-}
-
-.optionsChoice {
-    background-color: white;
-    border: none;
-    border-radius: 8px;
-    color: black;
-    padding-right: 10px;
-    padding-left: 10px;
-    height: 20px;
-    font-size: 12px;
-    cursor: pointer;
-    align-content: center;
-    text-align: center;
-}
-
-.optionsChoice:hover {
-    background-color: #fafafa; /* Grey color */
-}
-
-.activeButton {
-    background-color: #f0f0f0
-}
-
-
-
-
-
-
-
-
-.CodeSearch {
-    background-color: white; /* Very light grey background */
-    border: grey;
-    height: 22px;
-    width: 100px;
-    padding-left: 10px; /* Optional: adds padding inside the input */
-    margin-left: 10px; /* Adds space between the minus button and input */
-    outline: none; /* Ensure no outline when focused */
-    font-size: 12px; /* Set font size to 12px */
-    font-weight: 500; /* Set font weight to 500 */
-    border-radius: 8px; /* Rounded corners */
-}
-
-.CodeSearch:focus {
-    border: grey; /* Ensure no border when focused */
-    outline: none; /* Ensure no outline when focused */
-}
-
-.CodeSearch::placeholder {
-    color: grey;
-    font-style: italic; /* Make placeholder text italic */
-}
-
-.AddCourseBox {
-	display: flex;
-	flex-direction: row;
-	justify-content: space-between;
-	align-items: center;
-
-	width: 425px;
-	height: 36px;
-
-	border-radius: 16px;
-	margin-bottom: 5px;
-
-	padding-left: 10px;
-	padding-right: 10px;
-
-	background-color: #F5F5F5;
-	transition: filter 0.4s ease;
-}
-
-
-
-
-
-.TermSelect {
-  border: 1px solid white;
-  padding: 0 5px; /* Adjust padding to fit within the desired height */
-  background-color: white;
-  cursor: pointer;
-  border-radius: 4px;
-  font-size: 12px; /* Make the font smaller */
-  height: 22px; /* Set height */
-  display: flex; /* Align items within */
-  align-items: center; /* Center items vertically */
-  box-sizing: border-box; /* Include border and padding in the height */
-  color: black; /* Make text black */
-  position: relative; /* Make it the reference point for absolute positioning */
-}
-
-.TermDrop {
-  position: absolute;
-  background-color: white;
-  border: 1px solid white; /* Make the border white */
-  border-radius: 4px;
-  top: 100%; /* Position it directly below the TermSelect element */
-  margin-top: 5px; /* Additional spacing if needed */
-  z-index: 10;
-  width: 100px; /* Make the dropdown wider */
-  font-size: 12px; /* Make the font smaller */
-  box-sizing: border-box;
-}
-
-.TermDrop div {
-  padding: 5px;
-  cursor: pointer;
-  color: black; /* Make text black */
-}
-
-.TermDrop div:nth-child(odd) {
-  background-color: white; /* Lighter grey background */
-}
-
-.TermDrop div:nth-child(even) {
-  background-color: #f0f0f0; /* Lighter grey background */
-}
-
-.TermDrop div:hover {
-  background-color: #e0e0e0; /* Light grey on hover */
-}
-
-
-
-
-
-
-.RemoveButton {
-	display: flex;
-	justify-content: center;
-	align-items: center;
-	width: 16px; /* adjust size as needed */
-	height: 16px;
-	border-radius: 50%;
-	background-color: #ededed;
-	cursor: pointer;
-	margin-right: 5px;
-	/* optional border for better visibility */
-	/* border: 1px solid #C0C0C0; */
-}
-
-.RemoveButton:hover {
-	background-color: #D3D3D3; /* slightly darker grey on hover */
-}
-
