+++ frontend/src/pages/Courses/year/YearBox.module.css
@@ -1,0 +1,18 @@
+
+.row {
+	display: flex;
+	flex-direction: row;
+}
+
+.Grade {
+	font-weight: 600;
+	font-size: 25px;
+	margin-right: 10px;
+}
+
+.yearComponent {
+	display: flex;
+	flex-direction: column;
+	margin-bottom: 20px;
+	transition: transform 0.5s ease, height 0.5s ease, width 0.5s ease;
+}
