+++ frontend/src/pages/Courses/year/semester/course/CourseBox.module.css
@@ -1,0 +1,51 @@
+
+.row {
+	display: flex;
+	flex-direction: row;
+}
+
+
+.RemoveButton {
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	width: 16px; /* adjust size as needed */
+	height: 16px;
+	border-radius: 50%;
+	background-color: #ededed;
+	cursor: pointer;
+	margin-right: 5px;
+	/* optional border for better visibility */
+	/* border: 1px solid #C0C0C0; */
+}
+
+.RemoveButton:hover {
+	background-color: #D3D3D3; /* slightly darker grey on hover */
+}
+
+.checkmark {
+	justify-content: center;
+	text-align: center;
+
+	font-weight: 550;
+	margin-right: 2px;
+}
+
+.courseBox {
+	display: flex;
+	flex-direction: row;
+	justify-content: space-between;
+	align-items: center;
+
+	width: 425px;
+	height: 36px;
+
+	border-radius: 16px;
+	margin-bottom: 5px;
+
+	padding-left: 10px;
+	padding-right: 10px;
+
+	background-color: #F5F5F5;
+	transition: filter 0.4s ease;
+}
