+++ frontend/src/pages/Courses/year/semester/course/CourseBox.tsx
@@ -1,13 +1,13 @@
 
-import styles from "./../Courses.module.css";
+import Style from "./CourseBox.module.css";
 import "react-tooltip/dist/react-tooltip.css";
 
-import img_fall from "./../../../commons/images/fall.png";
-import img_spring from "./../../../commons/images/spring.png";
-import DistributionsCircle from "../../../commons/components/icons/DistributionsCircle"
+import img_fall from "./../../../../../commons/images/fall.png";
+import img_spring from "./../../../../../commons/images/spring.png";
+import DistributionsCircle from "./../../../../../commons/components/icons/DistributionsCircle"
 
-import { StudentCourse } from "./../../../commons/types/TypeCourse";
-import { User } from "../../../commons/types/TypeUser";
+import { StudentCourse } from "../../../../../commons/types/TypeCourse";
+import { User } from "../../../../../commons/types/TypeUser";
 // import { useModal } from "../../../hooks/modalContext";
 
 function RemoveCourse(props: { SC: StudentCourse, user: User, setUser: Function }){
@@ -56,7 +56,7 @@ function RemoveCourse(props: { SC: StudentCourse, user: User, setUser: Function 
   };
 
 	return( 
-		<div className={styles.RemoveButton} onClick={remove}>
+		<div className={Style.RemoveButton} onClick={remove}>
 			
 		</div>
 	)
@@ -74,32 +74,32 @@ function CourseBox(props: {edit: boolean, SC: StudentCourse, user: User, setUser
 		const renderMark = () => {
 			if(status === "DA_COMPLETE" || status === "DA_PROSPECT"){
 				return (
-					<div className={styles.checkmark}>
+					<div className={Style.checkmark}>
 						✓
 					</div>
 				);
 			}else if(status === "MA_HYPOTHETICAL" || "MA_VALID"){
 				const mark = (status === "MA_HYPOTHETICAL") ? "⚠" : "☑";
 				return (
-					<div className={styles.row}>
+					<div className={Style.row}>
 						{props.edit && <RemoveCourse SC={props.SC} user={props.user} setUser={props.setUser} />}
-						<div className={styles.checkmark}>
+						<div className={Style.checkmark}>
 							{mark}
 						</div>
 					</div>
 				);
 			}
-			return <div className={styles.hidden}></div>;
+			return <div></div>;
 		};
 	
 		const getBackgroundColor = () => (status === "DA_COMPLETE" ? "#E1E9F8" : "#F5F5F5");
 		const getSeasonImage = () => (String(term).endsWith("3") ? img_fall : img_spring);
 	
 		return (
-			<div className={styles.courseBox}  style={{ backgroundColor: getBackgroundColor() }}> 
+			<div className={Style.courseBox}  style={{ backgroundColor: getBackgroundColor() }}> 
 			{/* onClick={openModal} */}
 
-				<div className={styles.row} style={{ alignItems: "center" }}>
+				<div className={Style.row} style={{ alignItems: "center" }}>
 					{renderMark()}
 					<img style={{ width: "15px", height: "15px", marginRight: "6px" }} src={getSeasonImage()} alt="" />
 					<div>
@@ -112,7 +112,7 @@ function CourseBox(props: {edit: boolean, SC: StudentCourse, user: User, setUser
 					</div>
 				</div>
 				<div>
-					<div className={styles.row} style={{ alignItems: "center" }}>
+					<div className={Style.row} style={{ alignItems: "center" }}>
 						<DistributionsCircle distributions={[...course.areas, ...course.skills]} />
 					</div>
 				</div>
